# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SERVED. On-Demand Process Serving Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/subscription:
    get:
      tags: [marketplace]
      operationId: getMySubscription
      summary: Caller's ProServe subscription (attorney only)
      responses:
        '200':
          description: Subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '404':
          description: No subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /me/vault-subscription:
    get:
      tags: [marketplace]
      operationId: getMyVaultSubscription
      summary: Caller's Personal Vault subscription (requester only)
      responses:
        '200':
          description: Vault subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultSubscription'
        '404':
          description: No vault subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /me/credentialing:
    get:
      tags: [marketplace]
      operationId: getMyCredentialing
      summary: Caller's server credentialing status (server only)
      responses:
        '200':
          description: Credentialing record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerCredential'
        '404':
          description: Not started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /me/server-profile:
    get:
      tags: [marketplace]
      operationId: getMyServerProfile
      summary: Caller's server profile (server only)
      description: |
        Returns the caller's own server record. Used by the credentialing
        page to render the Server Profile panel (business address + Nevada
        process-server credentials).
      responses:
        '200':
          description: Server profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '404':
          description: No server record on file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [marketplace]
      operationId: updateMyServerProfile
      summary: Update the caller's own server profile (server only)
      description: |
        Lets a server self-edit profile fields needed for Nevada-compliant
        affidavits — business address, NV-licensed flag, license number,
        and license county. Server tier and active flag are not editable
        here.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMyServerProfileBody'
      responses:
        '200':
          description: Updated server profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: No server record on file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /me/payouts/instant:
    post:
      tags: [marketplace]
      operationId: triggerInstantPayout
      summary: |
        "Pay Now" — pull the connected account's available balance to the
        server's bank as an off-schedule standard payout (~2 business
        days). Does not affect the account's normal weekly schedule.
      responses:
        '200':
          description: Payout created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstantPayoutResponse'
        '400':
          description: |
            No Connect account, payouts not yet enabled, no available
            balance, or Stripe rejected the payout.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    Server:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        serverTier:
          type: string
          enum: [basic, pro, licensed]
        licenseNumber:
          type: string
        serviceArea:
          type: string
        businessAddress:
          type: string
          nullable: true
          description: Server's business / mailing address. Rendered into the affidavit attestation block.
        isLicensedNvServer:
          type: boolean
          description: Server attests they hold a Nevada PILB Process Server work card.
        licenseCounty:
          type: string
          nullable: true
          description: Nevada county of license registration (when isLicensedNvServer = true).
        serverType:
          type: string
          nullable: true
          enum: [licensed_nv, registered, private, sheriff]
          description: Self-declared server classification rendered on the affidavit attestation block.
        photoUrl:
          type: string
          nullable: true
          description: Object-storage path of the server's profile headshot (from /storage/profile-photo/request-url). Null until uploaded.
        active:
          type: boolean
        jobsCompleted:
          type: integer
        stripeAccountId:
          type: string
          nullable: true
          description: Stripe Connect Express account id (null until onboarding starts).
        payoutsEnabled:
          type: boolean
          description: True iff Stripe reports both charges_enabled && payouts_enabled.
        verifiedAt:
          type: string
          format: date-time
          nullable: true
          description: Mirrored from server_credentials.verified_at; populated after the background check passes.
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - serverTier
        - active
        - jobsCompleted
        - payoutsEnabled
        - createdAt
...
    UpdateMyServerProfileBody:
      type: object
      description: |
        Self-edit profile fields available to a server on the
        credentialing page. Mutates only Nevada-affidavit-relevant fields;
        Stripe state, server tier, and active flag are not editable here.
      properties:
        businessAddress:
          type: string
          description: Business address printed on Nevada affidavits.
        isLicensedNvServer:
          type: boolean
          description: |
            True if this server holds a Nevada PILB process-server work
            card. Drives the credentialing line that appears in the
            affidavit body.
        licenseNumber:
          type: string
        licenseCounty:
          type: string
          description: County of license — also used as the affidavit venue.
        serverType:
          type: string
          enum: [licensed_nv, registered, private, sheriff]
          description: Self-declared server classification (licensed_nv | registered | private | sheriff).
        photoUrl:
          type: string
          nullable: true
          description: Object-storage path of the server's profile headshot (from /storage/profile-photo/request-url). Send null to clear.
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Subscription:
      type: object
      required: [id, userId, tier, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        tier:
          type: string
          enum: [solo, firm, firm_pro]
        status:
          type: string
          enum: [active, past_due, canceled, incomplete]
        stripeSubscriptionId:
          type: string
          nullable: true
        stripeCustomerId:
          type: string
          nullable: true
        stripePriceId:
          type: string
          nullable: true
        currentPeriodEnd:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VaultSubscription:
      type: object
      required: [id, userId, tier, status, storageGb, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        tier:
          type: string
          enum: [basic, legal, pro]
        status:
          type: string
          enum: [active, past_due, canceled, incomplete]
        storageGb:
          type: integer
        stripeSubscriptionId:
          type: string
          nullable: true
        stripeCustomerId:
          type: string
          nullable: true
        stripePriceId:
          type: string
          nullable: true
        currentPeriodEnd:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ServerCredential:
      type: object
      required: [id, userId, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        status:
          type: string
          enum: [unpaid, pending, verified, failed]
        stripePaymentIntentId:
          type: string
          nullable: true
        paymentId:
          type: integer
          nullable: true
        backgroundCheckId:
          type: string
          nullable: true
        failureReason:
          type: string
          nullable: true
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    InstantPayoutResponse:
      type: object
      required: [payoutId, amountCents, status]
      properties:
        payoutId:
          type: string
          description: Stripe payout id (po_…) on the connected account.
        amountCents:
          type: integer
          description: Amount paid out in cents.
        arrivalDate:
          type: string
          format: date-time
          nullable: true
          description: Stripe's estimated bank arrival timestamp.
        status:
          type: string
          description: Stripe payout status (pending/in_transit/paid/failed/canceled).
...